Make sure rakubrew initializes correctly with broken versions
If there is a failed install, it's possible a not completely compiled
rakudo is lying around. In that case the init logic would fail, because
it can't determine the bin paths of that broken version to clean the $PATH
variable. Ignore such broken versions so it's at least possible to
initialize rakubrew properly.