Get `build` working on MacOS
Build in a temporary directory that has no space in it and move the
installation over to the versions dir afterwards.
This is hopefully a temporary hack, only required, because rakudo currently
can't build in folder with spaces in them. Once that is fixed this commit
can be reverted again. In any case older rakudos that don't even run in a
folder with spaces in it won't work in any case, as the versions dir always
has a space in it on MacOS. In addition for moving of the installation to
work the `--relocatable` option is necessary. That's also only available in
newer Rakudos.