Rip out the tmp dir building logic
I give up on making Rakudo able to be built in a folder with spaces for
now. Especially the external dependencies of MoarVM (dyncall, ...) are
difficult to do. e.g. dyncall has a build system that needs to work on a
much larger variety of platforms and make variants. Some of those
explicitly don't support escaping paths.
So I now aim to default to a folder that won't have a space in most cases:
- $HOME/.rakubrew on Linux and MacOS
- C:\rakubrew on Windows
And inform the user of the limitation and recommend to use $RAKUBREW_HOME
to set a different folder without spaces should the default one not work.