Linux fatpacking: trace HTTP::Tinyish backends
HTTP::Tinyish retrieval backends are dynamically detected and loaded.
Fatpack is not intelligent enough to trace that. So we include them in the
trace explicitly.
Also switch over to using --use for the Shell:: modules as well, as this is
more foolproof than just adding them manually to the trace list.