Fix `rakubrew triple` command
The `_get_git_cache_option` relied on cwd to be in a
rakudo checkout. That condition was not met by the call in
the `triple` implementation. Now the dir of the rakudo
checkout is taken explicitly as an argument in
`_get_git_cache_option`. That should help make introducing
such bugs in the future harder.