Be more clear about the `Sh` shell hook applying to all ~POSIX shells
This hook implementation restricts itself to features all compatible shells
should provide (or so I hope). So it can be used with all such shells.
Be more clear about that so people are not confused when they know they use
dash, or ksh or similar and rakubrew only talks about "the `sh` shell".

Also make the init command explicitly use the `Sh` hook. This is a bit less
magical than auto-detection, which I guess is a good thing.