Fix downloading when the latest compiled release <> latest release
When no version is given don't look at git for the latest release. This
fixes downloading when no version was given and git isn't installed. Also
don't rely on the 'latest' markers in the release index when no version was
given as the latest version might not have compiled release files yet.