Fix PAR changing env vars on MacOS
On MacOS Rakubrew is deployed via the PAR packager. On MacOS rakubrew
itself acts as the shim in `shim` mode. PAR adds environment variables and
modifies the dynamic library loading path variable. Those changes can cause
issues in the program being called. Thus we now try to revert the
environment variables to their original state before doing `exec` in the
shimming code.
Also add a test that compares the environment of a shimmed program to the
outer environment of the caller shell. This test is of limited use, as it
by default only works with the non PAR packaged Rakubrew. Thus the ability
to test a third party `rakubrew` executable specified via
`RAKUBREW_TEST_EXEC` is implemented.

This should hopefully fix #9.