Fix `download` on Windows
PAR doesn't know about DLL dependencies. One needs to explicitly tell it
which DLLs to include in the build. Without those DLLs HTTP::Tinyish
skipped LWP and HTTP::Tiny because those DLLs were missing for SSL support.
It then chose the Curl backend. That contained a bug, which messed up the
downloaded content by converting CRLF -> LF. That is now fixed in
https://github.com/miyagawa/HTTP-Tinyish/pull/24

We need to explicitly use PAR::Packer 1.057 and not the latest version,
because of another bug that causes Dynaloader to fail to find the included
DLLs and thus the user observing the same failure mode. See
https://github.com/rschupp/PAR-Packer/issues/86