Fix detection of x86 PowerShell
In general Powershell prepends the user profile path to $Env:PSMODULEPATH
when it starts up. But x86 PowerShells additionally prepend an x86 system
module path. So fix by not checking "starts with", but "contains".

Also this detection did and will continue to fail on Powershell version 1,
which didn't have a module system. But that's pretty much irrelevant.
Windows 7 was already shipped with version 2. So this only affects XP aera
systems.