Fix `build-zef` on Windows
It seems `&&` is not a thing in CMD. Running multiple `&&` joined commands
in a single `system()` seems to pass the entire thing literally to a single
program. In this case git errored out with an "unknown argument --head",
presumably because all the args, including the "--head" were passed to the
first git.